+++ packages/expo-audio/ios/AudioModule.swift
@@ -184,6 +184,25 @@ public class AudioModule: Module {
         }
       }
 
+      Function("setActiveForLockScreen") { (player: AudioPlayer, active: Bool, metadata: Metadata?) in
+        player.setActiveForLockScreen(active, metadata: metadata)
+      }
+
+      Function("updateLockScreenMetadata") { (player: AudioPlayer, metadata: Metadata?) in
+        if player.isActiveForLockScreen {
+          player.metadata = metadata
+          MediaController.shared.updateNowPlayingInfo(for: player)
+        }
+      }
+
+      Function("clearLockScreenControls") { (player: AudioPlayer) in
+        if player.isActiveForLockScreen {
+          player.metadata = nil
+          player.isActiveForLockScreen = false
+          MediaController.shared.setActivePlayer(nil)
+        }
+      }
+
       AsyncFunction("seekTo") { (player: AudioPlayer, seconds: Double) in
         await player.ref.currentItem?.seek(
           to: CMTime(
@@ -296,7 +315,7 @@ public class AudioModule: Module {
     try AudioUtils.validateAudioMode(mode: mode)
     let session = AVAudioSession.sharedInstance()
     var category: AVAudioSession.Category = session.category
-    var options: AVAudioSession.CategoryOptions = session.categoryOptions
+    var options: AVAudioSession.CategoryOptions = []
     self.shouldPlayInBackground = mode.shouldPlayInBackground
 
     #if os(iOS)
@@ -332,7 +351,11 @@ public class AudioModule: Module {
       }
     }
 
-    try AVAudioSession.sharedInstance().setCategory(category, options: options)
+    if options.isEmpty {
+      try AVAudioSession.sharedInstance().setCategory(category, mode: .default)
+    } else {
+      try AVAudioSession.sharedInstance().setCategory(category, options: options)
+    }
   }
 
   private func checkPermissions() throws {