+++ packages/expo-audio/ios/AudioPlayer.swift
@@ -11,6 +11,8 @@ public class AudioPlayer: SharedRef<AVPlayer> {
   var isLooping = false
   var shouldCorrectPitch = false
   var pitchCorrectionQuality: AVAudioTimePitchAlgorithm = .varispeed
+  var isActiveForLockScreen = false
+  var metadata: Metadata?
   var currentRate: Float = 0.0
   let interval: Double
   var wasPlaying = false
@@ -59,6 +61,10 @@ public class AudioPlayer: SharedRef<AVPlayer> {
     addPlaybackEndNotification()
     registerTimeObserver()
     ref.playImmediately(atRate: rate)
+
+    if isActiveForLockScreen {
+      MediaController.shared.updateNowPlayingInfo(for: self)
+    }
   }
 
   func setSamplingEnabled(enabled: Bool) {
@@ -93,12 +99,25 @@ public class AudioPlayer: SharedRef<AVPlayer> {
     ]
   }
 
+  func setActiveForLockScreen(_ active: Bool = true, metadata: Metadata? = nil) {
+    self.metadata = metadata
+    if active {
+      MediaController.shared.setActivePlayer(self)
+    } else {
+      MediaController.shared.setActivePlayer(nil)
+    }
+  }
+
   func updateStatus(with dict: [String: Any]) {
     var arguments = currentStatus()
     arguments.merge(dict) { _, new in
       new
     }
     self.emit(event: AudioConstants.playbackStatus, arguments: arguments)
+
+    if isActiveForLockScreen {
+      MediaController.shared.updateNowPlayingInfo(for: self)
+    }
   }
 
   private func setupPublisher() {
@@ -225,6 +244,11 @@ public class AudioPlayer: SharedRef<AVPlayer> {
   public override func sharedObjectWillRelease() {
     AudioComponentRegistry.shared.remove(self)
     setSamplingEnabled(enabled: false)
+
+    if isActiveForLockScreen {
+      MediaController.shared.setActivePlayer(nil)
+    }
+
     if let token = timeToken {
       ref.removeTimeObserver(token as Any)
     }